Move Gemma4 proportional RoPE to shared utilities
