+++ mlx_vlm/models/diffusion_gemma/language.py
@@ -13,8 +13,8 @@ from ..base import (
 from ..cache import KVCache, RotatingKVCache, StaticPrefixKVCache
 from ..gemma4.gemma4 import MultimodalEmbedder, masked_scatter
 from ..gemma4.language import RMSNormNoScale
-from ..gemma4.rope_utils import initialize_rope
 from ..gemma4.vision import VisionModel
+from ..rope_utils import initialize_rope
 from ..switch_layers import SwitchLinear, _gather_sort, _scatter_unsort
 from .config import ModelConfig, TextConfig
